+++ CHANGELOG.md
@@ -1,11 +1,31 @@
-<!-- markdownlint-disable MD024 -->
 # Change Log
 
-## 0.5.0 [unreleased]
+## 0.6.0 [unreleased]
 
 ### Bug Fixes
 
 1. [#86](https://github.com/InfluxCommunity/influxdb3-python/pull/86): Refactor to `timezone` specific `datetime` helpers to avoid use deprecated functions
+
+## 0.5.0 [2024-05-17]
+
+### Features
+
+1. [#88](https://github.com/InfluxCommunity/influxdb3-python/pull/88): Add support for named query parameters:
+   ```python
+   from influxdb_client_3 import InfluxDBClient3
+
+   with InfluxDBClient3(host="https://us-east-1-1.aws.cloud2.influxdata.com",
+                        token="my-token",
+                        database="my-database") as client:
+
+        table = client.query("select * from cpu where host=$host", query_parameters={"host": "server01"})
+
+        print(table.to_pandas())
+
+    ```
+
+### Bug Fixes
+
 1. [#87](https://github.com/InfluxCommunity/influxdb3-python/pull/87): Fix examples to use `write_options` instead of the object name `WriteOptions`
 
 ### Others
@@ -20,4 +40,4 @@
 
 ### Others
 
-- [#80](https://github.com/InfluxCommunity/influxdb3-python/pull/80): Integrate code style check into CI
+1. [#80](https://github.com/InfluxCommunity/influxdb3-python/pull/80): Integrate code style check into CI